policies: Move methods getRewriteRegexFromPathTemplate and getPathParamToIndexMap from common.go to route_utils.go
