+++ adapter/internal/oasparser/envoyconf/routes_utils.go
@@ -19,9 +19,13 @@ package envoyconf
 
 import (
 	"errors"
+	"fmt"
+	"regexp"
+	"strings"
 
 	routev3 "github.com/envoyproxy/go-control-plane/envoy/config/route/v3"
 	envoy_type_matcherv3 "github.com/envoyproxy/go-control-plane/envoy/type/matcher/v3"
+	logger "github.com/wso2/product-microgateway/adapter/internal/loggers"
 )
 
 func generateHTTPMethodMatcher(methodRegex string) []*routev3.HeaderMatcher {
@@ -61,6 +65,48 @@ func generateRegexMatchAndSubstitute(routePath, endpointBasePath,
 		},
 		Substitution: substitutionString,
 	}
+}
+
+// getRewriteRegexFromPathTemplate returns a regex with capture groups for given rewritePathTemplate
+// replaces {uri.var.petId} included in rewritePath given in policy
+// with indexes such as \1 \2 that are expected in the substitution string
+func getRewriteRegexFromPathTemplate(pathTemplate, rewritePathTemplate string) (string, error) {
+	rewriteRegex := "/" + strings.TrimSuffix(strings.TrimPrefix(rewritePathTemplate, "/"), "/")
+	pathParamToIndexMap := getPathParamToIndexMap(pathTemplate)
+	r := regexp.MustCompile(`{uri.var.([^{}]+)}`) // define a capture group to catch the path param
+	matches := r.FindAllStringSubmatch(rewritePathTemplate, -1)
+	for _, match := range matches {
+		// match is slice always with length two (since one capture group is defined in the regex)
+		// hence we do not want to explicitly validate the slice length
+		templatedParam := match[0]
+		param := match[1]
+		if index, ok := pathParamToIndexMap[param]; ok {
+			rewriteRegex = strings.ReplaceAll(rewriteRegex, templatedParam, fmt.Sprintf(`\%d`, index))
+		} else {
+			return "", fmt.Errorf("invalid path param %q in rewrite path", param)
+		}
+	}
+
+	// validate rewriteRegex
+	if matched, _ := regexp.MatchString(`^[a-zA-Z0-9~/_.\-\\]*$`, rewriteRegex); !matched {
+		logger.LoggerOasparser.Error("Rewrite path includes invalid characters")
+		return "", fmt.Errorf("rewrite path regex includes invalid characters, regex %q", rewriteRegex)
+	}
+
+	return rewriteRegex, nil
+}
+
+// getPathParamToIndexMap returns a map of path params to its index (map of path param -> index)
+func getPathParamToIndexMap(pathTemplate string) map[string]int {
+	indexMap := make(map[string]int)
+	r := regexp.MustCompile(`{([^{}]+)}`) // define a capture group to catch the path param
+	matches := r.FindAllStringSubmatch(pathTemplate, -1)
+	for i, paramMatches := range matches {
+		// paramMatches is slice always with length two (since one capture group is defined in the regex)
+		// hence we do not want to explicitly validate the slice length
+		indexMap[paramMatches[1]] = i + 1
+	}
+	return indexMap
 }
 
 func castPolicyParamMap(policyParams interface{}) (map[string]string, error) {
