+++ adapter/internal/oasparser/model/common.go
@@ -22,7 +22,6 @@ package model
 
 import (
 	"errors"
-	"fmt"
 	"net/url"
 	"regexp"
 	"strconv"
@@ -209,44 +208,4 @@ func unmarshalSwaggerResources(path string, methods []*Operation, vendorExtensio
 		//security:         operation.Security,
 		vendorExtensions: vendorExtensions,
 	}
-}
-
-// getRewriteRegexFromPathTemplate returns a regex with capture groups for given rewritePathTemplate
-func getRewriteRegexFromPathTemplate(pathTemplate, rewritePathTemplate string) (string, error) {
-	rewriteRegex := "/" + strings.TrimSuffix(strings.TrimPrefix(rewritePathTemplate, "/"), "/")
-	pathParamToIndexMap := getPathParamToIndexMap(pathTemplate)
-	r := regexp.MustCompile(`{uri.var.([^{}]+)}`) // define a capture group to catch the path param
-	matches := r.FindAllStringSubmatch(rewritePathTemplate, -1)
-	for _, match := range matches {
-		// match is slice always with length two (since one capture group is defined in the regex)
-		// hence we do not want to explicitly validate the slice length
-		templatedParam := match[0]
-		param := match[1]
-		if index, ok := pathParamToIndexMap[param]; ok {
-			rewriteRegex = strings.ReplaceAll(rewriteRegex, templatedParam, fmt.Sprintf(`\%d`, index))
-		} else {
-			return "", fmt.Errorf("invalid path param %q in rewrite path", param)
-		}
-	}
-
-	// validate rewriteRegex
-	if matched, _ := regexp.MatchString(`^[a-zA-Z0-9~/_.\-\\]*$`, rewriteRegex); !matched {
-		logger.LoggerOasparser.Error("Rewrite path includes invalid characters")
-		return "", fmt.Errorf("rewrite path regex includes invalid characters, regex %q", rewriteRegex)
-	}
-
-	return rewriteRegex, nil
-}
-
-// getPathParamToIndexMap returns a map of path params to its index (map of path param -> index)
-func getPathParamToIndexMap(pathTemplate string) map[string]int {
-	indexMap := make(map[string]int)
-	r := regexp.MustCompile(`{([^{}]+)}`) // define a capture group to catch the path param
-	matches := r.FindAllStringSubmatch(pathTemplate, -1)
-	for i, paramMatches := range matches {
-		// paramMatches is slice always with length two (since one capture group is defined in the regex)
-		// hence we do not want to explicitly validate the slice length
-		indexMap[paramMatches[1]] = i + 1
-	}
-	return indexMap
 }
